api_backend/requirements.txt
# Framework Web e Servidor ASGI
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Banco de Dados (MongoDB)
pymongo[srv]==4.8.0

# Validação de Dados e Configurações
pydantic==2.7.4
pydantic-settings==2.3.4

# Carregamento de variáveis de ambiente
python-dotenv==1.0.1

# Integração com o Google Gemini
google-generativeai==0.7.0

# Integração com o Telegram
python-telegram-bot==21.0.1

# Bibliotecas para o Simulador
requests==2.31.0

# --- NOVO: Bibliotecas para o Dashboard (Streamlit) ---
streamlit
pandas
plotly